md5.cpp: shut up 64-bit MSC warning.
git-svn-id: http://www.virtualbox.org/svn/vbox/trunk@14048 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Runtime/common/checksum/md5.cpp
@@ -111,7 +111,7 @@ RTDECL(void) RTMd5Update(PRTMD5CONTEXT ctx, const void *pvBuf, size_t len)
     t = ctx->bits[0];
     if ((ctx->bits[0] = t + ((uint32_t) len << 3)) < t)
 	ctx->bits[1]++; 	/* Carry from low to high */
-    ctx->bits[1] += len >> 29;
+    ctx->bits[1] += (uint32_t)(len >> 29);
 
     t = (t >> 3) & 0x3f;	/* Bytes already in shsInfo->data */
